WIP: thumbnail refactor
