+++ app/assets/stylesheets/application.scss
@@ -13,7 +13,7 @@
 @import 'simple_form';
 @import 'communities';
 @import 'items';
-@import 'logo';
+@import 'image';
 @import 'search';
 @import 'users';
 @import 'item_draft';
