+++ app/assets/stylesheets/image.scss
@@ -1,0 +1,9 @@
+.j-thumbnail {
+  height: 100px;
+  width: 100px;
+
+  img {
+    max-height: 100px;
+    max-width: 100px;
+  }
+}
