--- app/assets/stylesheets/logo.scss
@@ -1,9 +1,0 @@
-.logo-large {
-  height: 400px;
-}
-
-.logo-small {
-  height: 100px;
-  max-width: inherit;
-  width: 100px;
-}
