+++ app/views/admin/communities/_form.html.erb
@@ -1,6 +1,6 @@
 <%= simple_form_for [:admin, @community] do |f| %>
   <div class="form-group">
-    <%= render partial: 'logo', locals: {object: @community} %>
+    <%= render partial: 'thumbnail', locals: { thumbnail: @community.logo.attachment } %>
 
     <%= f.input :logo, as: :file, label: false, wrapper_html: { class: 'mt-3' } %>
