+++ app/views/application/_footer.html.erb
@@ -10,7 +10,7 @@
                  '#TODO', class: 'text-dark small' %></p>
       </div>
       <div class="order-first p-1 py-3">
-        <%= image_tag 'ualib-logo.png', alt: 'logo' %>
+        <%= image_tag 'ualib-logo.png', alt: t('.uofa_library_logo_alt')%>
       </div>
     </div>
   </div>
