+++ app/views/communities/show.html.erb
@@ -1,8 +1,7 @@
 <% page_title(@community.title) %>
 
-<%= render partial: 'show_breadcrumbs' %>
-
 <div class="container mt-3">
+  <%= render partial: 'show_breadcrumbs' %>
 
   <div class="d-flex flex-wrap align-items-center justify-content-between mt-3">
     <h1><%= @community.title %></h1>
@@ -29,8 +28,8 @@
   </div>
 
   <div class="media p-3">
-    <%= render partial: 'logo', locals: {object: @community} %>
-    <div class="media-body">
+    <%= render partial: 'thumbnail', locals: { thumbnail: @community.logo.attachment } %>
+    <div class="media-body ml-3">
       <p><%= @community.description %></p>
     </div>
   </div>
