+++ app/views/welcome/index.html.erb
@@ -1,4 +1,4 @@
-<%= image_tag 'campus-banner.jpg', class: 'img-fluid w-100', alt: 'banner image' %>
+<%= image_tag 'campus-banner.jpg', class: 'img-fluid w-100', alt: '' %>
 
 <section class="jumbotron text-center">
   <div class="container">
@@ -43,7 +43,7 @@
     <div class="d-flex flex-row">
       <div class="j-image-with-text mx-3 col-md-3">
         <%= link_to '#TODO' do %>
-          <%= image_tag 'dept-of-agriculture.jpg', alt: t('.afns_collection'),
+          <%= image_tag 'dept-of-agriculture.jpg', alt: '',
               class: 'agriculture' %>
           <h4 class="small font-weight-bold"><%= t('.afns_collection') %></h4>
         <% end %>
@@ -54,7 +54,7 @@
         <div class="d-flex flex-row mb-3">
           <div class="j-image-with-text mr-3">
             <%= link_to '#TODO' do %>
-              <%= image_tag 'images-of-research.jpg', alt: t('.images_of_research_collection'),
+              <%= image_tag 'images-of-research.jpg', alt: '',
                   title: t('.images_of_research_attribution') %>
               <h4 class="small font-weight-bold"><%= t('.images_of_research_collection') %></h4>
             <% end %>
@@ -62,7 +62,7 @@
 
           <div class="j-image-with-text">
             <%= link_to '#TODO' do %>
-              <%= image_tag 'theses.jpg', alt: t('.theses_collection') %>
+              <%= image_tag 'theses.jpg', alt: '' %>
               <h4 class="small font-weight-bold"><%= t('.theses_collection') %></h4>
             <% end %>
           </div>
@@ -71,7 +71,7 @@
 
         <div class="j-image-with-text mt-auto">
           <%= link_to '#TODO' do %>
-            <%= image_tag 'oil-sands.jpg', alt: t('.osrin_collection'),
+            <%= image_tag 'oil-sands.jpg', alt: '',
                 title: t('.osrin_attribution') %>
             <h4 class="small font-weight-bold"><%= t('.osrin_collection') %></h4>
           <% end %>
