+++ config/locales/en.yml
@@ -141,7 +141,7 @@ en:
 
   application:
     navbar:
-      logo_alt: 'ERA Logo'
+      era_logo_alt: 'ERA - Education and Research Archive'
       search:
         placeholder: 'Search ERA...'
         button_text: 'Search'
@@ -157,6 +157,7 @@ en:
       user_dropdown:
         heading_html: "Signed in as <strong>%{name}</strong>"
     footer:
+      uofa_library_logo_alt: 'University of Alberta Libraries'
       street_address: 'University of Alberta, Edmonton, AB, Canada, 26G 2R3'
       privacy_policy: 'Web Site Privacy Policy'
   locked_ldp_object:
@@ -536,8 +537,6 @@ en:
         no_saved_files: No files have been attached to this item
         item_thumbnail: 'Item Thumbnail'
         thumbnail_of_first_file: 'By default the first file will be displayed as the thumbnail image for this item. To change this, click on set as thumbnail button on the desired file.'
-      thumbnail:
-        thumbnail_alt: Thumbnail Image
       file:
         set_as_thumbnail: Set as Thumbnail
         delete_file_dialog: Are you sure you want to delete this file?
